pyproject.toml
[tool.poetry]
name = "land-fund-contracts"
version = "0.1.0"
description = ""
authors = ["Marko Culinovic <marko@high.rs>"]
readme = "README.md"
packages = [{include = "land_fund_contracts"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
eth-brownie = "^1.17.2"


[tool.poetry.group.dev.dependencies]
black = "^21.11b1"
flake8 = "^4.0.1"
isort = "^5.10.1"
pytest = "^6.2.5"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
